groupBaseclaim

+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Services/BackOffice/UiPageMetadataService.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -14,13 +16,16 @@ namespace TestingAndCalibrationLabs.Business.Services
     /// </summary>
     public class UiPageMetadataService : IUiPageMetadataService
     {
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        private readonly IAuthorizationService _authorizationService;
         private readonly IUiPageMetadataCharacteristicsRepository _uiPageMetadataCharacteristicsRepository;
         public readonly IUiPageMetadataRepository _uiPageMetadataRepository;
-        public UiPageMetadataService(IUiPageMetadataRepository uiPageMetadataRepository,IGenericRepository<UiPageMetadataModel> genericRepository, IUiPageMetadataCharacteristicsRepository uiPageMetadataCharacteristicsRepository)
+        public UiPageMetadataService(IHttpContextAccessor httpContextAccessor,IAuthorizationService authorizationService,IUiPageMetadataRepository uiPageMetadataRepository,IGenericRepository<UiPageMetadataModel> genericRepository, IUiPageMetadataCharacteristicsRepository uiPageMetadataCharacteristicsRepository)
         {
             _uiPageMetadataRepository = uiPageMetadataRepository;
             _uiPageMetadataCharacteristicsRepository = uiPageMetadataCharacteristicsRepository;
-           
+           _httpContextAccessor= httpContextAccessor;
+            _authorizationService= authorizationService;
         }
         /// <summary>
         /// Insert Record In Ui Page Metadata Type
@@ -29,8 +34,12 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public RequestResult<int> Create(UiPageMetadataModel pageControl)
         {
-            int id = _uiPageMetadataRepository.Create(pageControl);
-            return new RequestResult<int>(1);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, pageControl, new[] { Operations.Create }).Result.Succeeded)
+            {
+                int id = _uiPageMetadataRepository.Create(pageControl);
+                return new RequestResult<int>(1);
+            }
+            return new RequestResult<int>(0);   
         }
         /// <summary>
         /// Delete Record From Ui Page Metadata Type
@@ -39,7 +48,11 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public bool Delete(int id)
         {
-            return _uiPageMetadataRepository.Delete(id);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiPageMetadataModel(), new[] { Operations.Delete }).Result.Succeeded)
+            {
+                return _uiPageMetadataRepository.Delete(id);
+            }
+            return false;
         }
         /// <summary>
         /// Get Record by Id For Ui Page Metadata Type
@@ -48,8 +61,11 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public UiPageMetadataModel GetById(int id)
         {
-
-            return _uiPageMetadataRepository.GetById(id);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiPageMetadataModel(), new[] { Operations.Read }).Result.Succeeded)
+            {
+                return _uiPageMetadataRepository.GetById(id);
+            }
+            return null;
         }
         /// <summary>
         /// Edit Record From Ui Page Metadata Type
@@ -59,8 +75,12 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public RequestResult<int> Update(int id, UiPageMetadataModel uiPageMetadataModel)
         {
-            _uiPageMetadataRepository.Update(uiPageMetadataModel);
-            return new RequestResult<int>(1);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, uiPageMetadataModel, new[] { Operations.Update }).Result.Succeeded)
+            {
+                _uiPageMetadataRepository.Update(uiPageMetadataModel);
+                return new RequestResult<int>(1);
+            }
+            return new RequestResult<int>(0);
         }
         /// <summary>
         /// Get All Records From Ui Page Metadata Type
@@ -68,7 +88,11 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public List<UiPageMetadataModel> Get()
         {
-            return _uiPageMetadataRepository.Get();
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiPageMetadataModel(), new[] { Operations.Read }).Result.Succeeded)
+            {
+                return _uiPageMetadataRepository.Get();
+            }
+            return null;
         }
     }
 }
